+++ core/auth/interfaces.go
@@ -305,6 +305,9 @@ type APIs interface {
 
 	//DeleteAPIKey deletes an API key
 	DeleteAPIKey(ID string) error
+
+	//ValidateAPIKey validates the given API key for the given app ID
+	ValidateAPIKey(appID string, apiKey string) error
 }
 
 //Storage interface to communicate with the storage
@@ -385,11 +388,11 @@ type Storage interface {
 	DeleteAPIKey(ID string) error
 
 	//ApplicationTypes
-	FindApplicationTypeByIdentifier(identifier string) (*model.ApplicationType, error)
+	FindApplicationType(id string) (*model.ApplicationType, error)
 
 	//ApplicationsOrganizations
 	LoadApplicationsOrganizations() ([]model.ApplicationOrganization, error)
-	FindApplicationOrganizations(appID string, orgID string) (*model.ApplicationOrganization, error)
+	FindApplicationOrganization(appID string, orgID string) (*model.ApplicationOrganization, error)
 
 	//Device
 	FindDevice(context storage.TransactionContext, deviceID string, accountID string) (*model.Device, error)