+++ CHANGELOG.md
@@ -11,12 +11,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Expose delete app org role admin API [#313](https://github.com/rokwire/core-building-block/issues/313)
 - Expose delete app org group admin API [#312](https://github.com/rokwire/core-building-block/issues/312)
 - Expose Admin API with the currently logged in accounts [#355](https://github.com/rokwire/core-building-block/issues/355)
+- Add app config endpoints [#261](https://github.com/rokwire/core-building-block/issues/261)
 
 ### Fixed
 - Fix creating application group admin API [#397](https://github.com/rokwire/core-building-block/issues/397)
 - Fix create role and create group Admin APIs [#386](https://github.com/rokwire/core-building-block/issues/386)
+- Fix broken app config API and CHANGELOG.md [#401](https://github.com/rokwire/core-building-block/issues/401)
 
 ### Changed
+- Limit the returned items for get application accounts admin API [#375](https://github.com/rokwire/core-building-block/issues/375)
 - Return ordered list of account auth types on link account auth type [#376](https://github.com/rokwire/core-building-block/issues/376)
 
 ## [1.18.0] - 2022-01-25